Moves tests to a separate file Fixes test issues
