+++ tests/integration/test_integration.py
@@ -8,7 +8,6 @@ import shutil
 import time
 from os.path import abspath
 from pathlib import Path
-from typing import List
 
 import hvac  # type: ignore[import-untyped]
 import pytest
@@ -24,7 +23,6 @@ METADATA = yaml.safe_load(Path("./metadata.yaml").read_text())
 
 APPLICATION_NAME = "vault-k8s"
 PROMETHEUS_APPLICATION_NAME = "prometheus-k8s"
-S3_INTEGRATOR_APPLICATION_NAME = "s3-integrator"
 TRAEFIK_APPLICATION_NAME = "traefik"
 SELF_SIGNED_CERTIFICATES_APPLICATION_NAME = "self-signed-certificates"
 VAULT_KV_REQUIRER_APPLICATION_NAME = "vault-kv-requirer"
@@ -131,105 +129,6 @@ class TestVaultK8s:
         )
         return await ops_test.model.get_action_output(action_uuid=action.entity_id, wait=timeout)
 
-    async def run_s3_integrator_sync_credentials_action(
-        self,
-        ops_test: OpsTest,
-    ) -> dict:
-        """Runs `sync-s3-credentials` action on the `s3-integrator` leader unit.
-
-        Args:
-            ops_test (OpsTest): OpsTest
-
-        Returns:
-            dict: Action output
-        """
-        assert ops_test.model
-        leader_unit = await get_leader_unit(ops_test.model, S3_INTEGRATOR_APPLICATION_NAME)
-        access_key = os.getenv("S3_ACCESS_KEY")
-        secret_key = os.getenv("S3_SECRET_KEY")
-        sync_credentials_action = await leader_unit.run_action(
-            action_name="sync-s3-credentials",
-            **{
-                "access-key": access_key,
-                "secret-key": secret_key,
-            },
-        )
-        return await ops_test.model.get_action_output(
-            action_uuid=sync_credentials_action.entity_id, wait=120
-        )
-
-    async def run_create_backup_action(
-        self,
-        ops_test: OpsTest,
-    ) -> dict:
-        """Runs `create-backup` action on the `vault-k8s` leader unit.
-
-        Args:
-            ops_test (OpsTest): OpsTest
-
-        Returns:
-            dict: Action output
-        """
-        assert ops_test.model
-        leader_unit = await get_leader_unit(ops_test.model, APPLICATION_NAME)
-        create_backup_action = await leader_unit.run_action(
-            action_name="create-backup",
-        )
-        return await ops_test.model.get_action_output(
-            action_uuid=create_backup_action.entity_id, wait=120
-        )
-
-    async def run_list_backups_action(
-        self,
-        ops_test: OpsTest,
-    ) -> dict:
-        """Runs `list-backups` action on the `vault-k8s` leader unit.
-
-        Args:
-            ops_test (OpsTest): OpsTest
-
-        Returns:
-            dict: Action output
-        """
-        assert ops_test.model
-        leader_unit = await get_leader_unit(ops_test.model, APPLICATION_NAME)
-        list_backups_action = await leader_unit.run_action(
-            action_name="list-backups",
-        )
-        return await ops_test.model.get_action_output(
-            action_uuid=list_backups_action.entity_id, wait=120
-        )
-
-    async def run_restore_backup_action(
-        self,
-        ops_test: OpsTest,
-        backup_id: str,
-        root_token: str,
-        unseal_keys: List[str],
-    ) -> dict:
-        """Runs `restore-backup` action on the `vault-k8s` leader unit.
-
-        Args:
-            ops_test (OpsTest): OpsTest
-
-        Returns:
-            dict: Action output
-        """
-        assert ops_test.model
-        leader_unit = await get_leader_unit(ops_test.model, APPLICATION_NAME)
-        restore_backup_action = await leader_unit.run_action(
-            action_name="restore-backup",
-            **{
-                "backup-id": backup_id,
-                "unseal-keys": unseal_keys,
-                "root-token": root_token,
-            },
-        )
-        restore_backup_action_output = await ops_test.model.get_action_output(
-            action_uuid=restore_backup_action.entity_id, wait=120
-        )
-        return restore_backup_action_output
-
     @pytest.mark.abort_on_fail
     @pytest.fixture(scope="module")
     async def deploy_prometheus(self, ops_test: OpsTest) -> None:
@@ -244,34 +143,6 @@ class TestVaultK8s:
             application_name=PROMETHEUS_APPLICATION_NAME,
             trust=True,
         )
-
-    @pytest.mark.abort_on_fail
-    @pytest.fixture(scope="module")
-    async def deploy_and_configure_s3_integrator(self, ops_test: OpsTest) -> None:
-        """Deploys S3 Integrator.
-
-        Args:
-            ops_test: Ops test Framework.
-        """
-        assert ops_test.model
-        await ops_test.model.deploy(
-            "s3-integrator",
-            application_name=S3_INTEGRATOR_APPLICATION_NAME,
-            trust=True,
-        )
-        s3_integrator = ops_test.model.applications[S3_INTEGRATOR_APPLICATION_NAME]
-        await ops_test.model.wait_for_idle(
-            apps=[S3_INTEGRATOR_APPLICATION_NAME],
-            status="blocked",
-            timeout=1000,
-            wait_for_exact_units=1,
-        )
-        await self.run_s3_integrator_sync_credentials_action(ops_test)
-        s3_config = {
-            "endpoint": os.getenv("S3_ENDPOINT"),
-            "bucket": "test-bucket",
-        }
-        await s3_integrator.set_config(s3_config)
 
     @pytest.mark.abort_on_fail
     @pytest.fixture(scope="module")
@@ -323,24 +194,6 @@ class TestVaultK8s:
         self, ops_test: OpsTest, build_and_deploy
     ):
         assert ops_test.model
-        await ops_test.model.wait_for_idle(
-            apps=[APPLICATION_NAME],
-            status="active",
-            timeout=1000,
-            wait_for_exact_units=NUM_VAULT_UNITS,
-        )
-
-    @pytest.mark.abort_on_fail
-    async def test_given_application_is_deployed_when_pod_crashes_then_unit_recovers(
-        self,
-        ops_test: OpsTest,
-        build_and_deploy,
-    ):
-        assert ops_test.model
-        unit = ops_test.model.units[f"{APPLICATION_NAME}/1"]
-        assert isinstance(unit, Unit)
-        k8s_namespace = ops_test.model.name
-        crash_pod(name=f"{APPLICATION_NAME}-1", namespace=k8s_namespace)
         await ops_test.model.wait_for_idle(
             apps=[APPLICATION_NAME],
             status="active",
@@ -400,7 +253,6 @@ class TestVaultK8s:
         self,
         ops_test: OpsTest,
     ):
-        """This proves that vault is reachable behind ingress."""
         vault_endpoint = await self._get_vault_endpoint(ops_test)
         action_output = await self.run_get_ca_certificate_action(ops_test)
         ca_certificate = action_output["ca-certificate"]
@@ -511,99 +363,14 @@ class TestVaultK8s:
         build_and_deploy,
     ):
         assert ops_test.model
-        num_units = 3
 
         app = ops_test.model.applications[APPLICATION_NAME]
         assert isinstance(app, Application)
-        await app.scale(num_units)
+        await app.scale(NUM_VAULT_UNITS)
 
         await ops_test.model.wait_for_idle(
             apps=[APPLICATION_NAME],
             status="active",
             timeout=1000,
-            wait_for_exact_units=num_units,
+            wait_for_exact_units=NUM_VAULT_UNITS,
         )
-
-    @pytest.mark.abort_on_fail
-    async def test_given_application_is_deployed_and_related_to_s3_integrator_when_create_backup_action_then_backup_is_created(
-        self,
-        ops_test: OpsTest,
-        build_and_deploy,
-        deploy_and_configure_s3_integrator,
-    ):
-        assert ops_test.model
-        await ops_test.model.integrate(
-            relation1=APPLICATION_NAME,
-            relation2=S3_INTEGRATOR_APPLICATION_NAME,
-        )
-        await ops_test.model.wait_for_idle(
-            apps=[S3_INTEGRATOR_APPLICATION_NAME],
-            status="active",
-            timeout=1000,
-        )
-        await ops_test.model.wait_for_idle(
-            apps=[APPLICATION_NAME],
-            status="active",
-            timeout=1000,
-            wait_for_exact_units=5,
-        )
-        vault = ops_test.model.applications[APPLICATION_NAME]
-        assert isinstance(vault, Application)
-        create_backup_action_output = await self.run_create_backup_action(ops_test)
-        assert create_backup_action_output["backup-id"]
-    
-    @pytest.mark.abort_on_fail
-    async def test_given_application_is_deployed_and_backup_created_when_list_backups_action_then_backups_are_listed(
-        self,
-        ops_test: OpsTest,
-        build_and_deploy,
-        deploy_and_configure_s3_integrator,
-    ):
-        assert ops_test.model
-        await ops_test.model.wait_for_idle(
-            apps=[S3_INTEGRATOR_APPLICATION_NAME],
-            status="active",
-            timeout=1000,
-        )
-        await ops_test.model.wait_for_idle(
-            apps=[APPLICATION_NAME],
-            status="active",
-            timeout=1000,
-            wait_for_exact_units=5,
-        )
-        vault = ops_test.model.applications[APPLICATION_NAME]
-        assert isinstance(vault, Application)
-        list_backups_action_output = await self.run_list_backups_action(ops_test)
-        assert list_backups_action_output["backup-ids"]
-
-    @pytest.mark.abort_on_fail
-    async def test_given_application_is_deployed_and_backup_created_when_restore_backup_action_then_backup_is_restored(
-        self,
-        ops_test: OpsTest,
-        build_and_deploy,
-        deploy_and_configure_s3_integrator,
-    ):
-        assert ops_test.model
-        await ops_test.model.wait_for_idle(
-            apps=[S3_INTEGRATOR_APPLICATION_NAME],
-            status="active",
-            timeout=1000,
-        )
-        await ops_test.model.wait_for_idle(
-            apps=[APPLICATION_NAME],
-            status="active",
-            timeout=1000,
-            wait_for_exact_units=5,
-        )
-        vault = ops_test.model.applications[APPLICATION_NAME]
-        assert isinstance(vault, Application)
-        list_backups_action_output = await self.run_list_backups_action(ops_test)
-        backup_id = json.loads(list_backups_action_output["backup-ids"])[0]
-        # In this test we are not using the correct unsealed keys and root token.
-        restore_backup_action_output = await self.run_restore_backup_action(
-            ops_test,
-            backup_id=backup_id,
-            root_token="RandomRootToken",
-            unseal_keys=["RandomUnsealKey"],
-        )
-        assert restore_backup_action_output["restored"] == backup_id
